pyproject.toml
[project]
name = "clearpixai"
version = "0.1.0"
description = "Simple watermark removal tool using AI"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "easyocr>=1.7.0",
    "transformers>=4.40.0",
    "diffusers>=0.27.0",
    "accelerate>=0.20.0",
    "safetensors>=0.4.0",
    "supervision>=0.21.0",
    "timm>=1.0.0",
    "einops>=0.6.0",
    "huggingface-hub>=0.24.0",
]

[project.scripts]
clearpixai = "clearpixai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


